Project script files
See README for description. Additional data files are uploaded to the ICL dropbox, but all data files are available on request (some of these contain licensed data with further use dependent on citation, acknowledgement or informing the authors). The Norberg bakeoff files (original version and my adapted version) are not listed here: they are uploaded to the ICL dropbox with input csv files and correct folder structure.